Consertando o bug no gráfico quando não havia agendamento no mês.

+++ src/main/java/com/uhc/quatropatas/dto/AgendamentoMes.java
@@ -6,12 +6,12 @@ package com.uhc.quatropatas.dto;
 public class AgendamentoMes {
 
 	private String mes;
-	private Integer total;
+	private Long total;
 
 	public AgendamentoMes(){
 	}
 	
-	public AgendamentoMes(String mes, Integer total){
+	public AgendamentoMes(String mes, Long total){
 	    this.mes = mes;
 	    this.total = total;
 	}
@@ -24,12 +24,20 @@ public class AgendamentoMes {
 		this.mes = mes;
 	}
 
-	public Integer getTotal() {
+	public Long getTotal() {
 		return total;
 	}
 
-	public void setTotal(Integer total) {
+	public void setTotal(Long total) {
 		this.total = total;
 	}
+
+//	public Integer getTotal() {
+//		return total;
+//	}
+//
+//	public void setTotal(Integer total) {
+//		this.total = total;
+//	}
 	
 }
